[#206] Moving comparers to more domain specific folder
