+++ src/FluentSerializer.Core.Tests/Tests/Context/SerializerCoreContextTests.cs
@@ -1,6 +1,5 @@
 using FluentAssertions;
 
-using FluentSerializer.Core.Comparing;
 using FluentSerializer.Core.Configuration;
 using FluentSerializer.Core.Context;
 using FluentSerializer.Core.DataNodes;
