+++ src/FluentSerializer.Core.Tests/Tests/DataNodes/DataNodeComparerTests.Equals.cs
