+++ src/FluentSerializer.Core.Tests/Tests/DataNodes/DataNodeComparerTests.GetHashCodeForAll.cs
