+++ src/FluentSerializer.Core.Tests/Tests/DataNodes/DataNodeComparerTests.cs
@@ -1,4 +1,3 @@
-using FluentSerializer.Core.Comparing;
 using FluentSerializer.Core.DataNodes;
 using FluentSerializer.Core.Text;
